forgi sampling ready, not yet debugged
