+++ graphlearn/abstract_graphs/directedgraphtools.py
@@ -33,3 +33,24 @@ def get_start_and_end_node(graph):
     if start==-1 or end==-1:
         raise Exception ('your beautiful "rna" has no clear start or end')
     return start,end
+
+
+
+
+
+def expanded_rna_graph_to_digraph(graph):
+    '''
+    :param graph:  an expanded rna representing graph as produced by eden.
+                   properties: backbone edges are replaced by a node labeled '-'.
+                   rna reading direction is reflected by ascending node ids in the graph.
+    :return: a graph, directed edges along the backbone
+    '''
+    digraph=nx.DiGraph(graph)
+    for n,d in digraph.nodes(data=True):
+        if 'edge' in d:
+            if d['label']=='-':
+                ns=digraph.neighbors(n)
+                ns.sort()
+                digraph.remove_edge(ns[1],n)
+                digraph.remove_edge(n,ns[0])
+    return digraph
